requirements.txt
dataset==0.6.4
python-dotenv==0.5.1
requests==2.10.0
slacker==0.9.16
xmltodict==0.10.2
